Move to a behavior based test approach
